user-clustering/api_cluster/requirements.txt
Flask
pandas
joblib
sqlalchemy
scikit-learn==1.4.2
gunicorn
pymysql
mysql-connector-python
pymysql
requests